// src/types.ts
export enum UserRole {
  SHIPPER_EXPORTER = 'Shipper / Exporter',
  LINER_OPERATOR = 'Ship Liner / Operator',
  FORWARDER_AGENT = 'Forwarder / Agent'
}

export type SubscriptionPlan = 'FREE' | 'OFFICE' | 'CORPORATE';

export interface UserProfile {
  email: string;
  mobile: string;
  role: UserRole;
  fullName: string;
  designation: string;
  companyName: string;
  address: string;
  photoBase64: string | null;
  logoBase64: string | null;
  onboardingAnalysis?: string;
  isMainUser?: boolean;
  branchId?: string;
  status?: 'ACTIVE' | 'SUSPENDED';
  plan: SubscriptionPlan;
  subscriptionStatus: 'ACTIVE' | 'EXPIRED';
  registrationDate: string;
  expiryDate: string;
}
// ...
export type AuthState = 'LOGIN_EMAIL' | 'LOGIN_PASS' | 'LOGIN_OTP' | 'SIGNUP' | 'DASHBOARD' | 'ADMIN_PANEL' | 'SUBSCRIPTION_EXPIRED';